nifi_chat_ui/requirements.txt
streamlit
openai
google-generativeai
python-dotenv
mcp 